Move results processing logic to a new pkg
plus fixes
